requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.2.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
discord.py==2.3.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gradio==4.29.0
gradio_client==0.16.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyboard==0.13.5
kiwisolver==1.4.5
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python==4.9.0.80
openpyxl==3.1.2
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pefile==2023.2.7
pillow==10.2.0
proglog==0.1.10
PyAutoGUI==0.9.54
pydantic==2.6.2
pydantic_core==2.16.3
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytweening==1.2.0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
ruff==0.2.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
tiktoken==0.6.0
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
websockets==11.0.3
Werkzeug==3.0.1
xlrd==2.0.1
yarl==1.9.4
